Oppdaterte readme for oppgave 2, dro ut objekter for oppgave ut i server-fila
